Refactor metrics configuration
