+++ src/anomalib/models/components/base/anomaly_module.py
@@ -10,7 +10,6 @@ from abc import ABC
 from typing import Any
 
 import lightning.pytorch as pl
-import torch
 from lightning.pytorch.callbacks import Callback
 from lightning.pytorch.utilities.types import STEP_OUTPUT
 from torch import Tensor, nn
@@ -94,65 +93,3 @@ class AnomalyModule(pl.LightningModule, ABC):
         del args, kwargs  # These variables are not used.
 
         return self.predict_step(batch, batch_idx)
-
-    def on_validation_batch_end(self, outputs, batch, batch_idx, dataloader_idx=0):
-        """Called at the end of each validation step."""
-        self._outputs_to_cpu(outputs)
-
-        self._collect_output(self.image_metrics, self.pixel_metrics, outputs)
-
-    def on_test_batch_end(self, outputs, batch, batch_idx, dataloader_idx=0):
-        """Called at the end of each test step."""
-        self._outputs_to_cpu(outputs)
-
-        self._collect_output(self.image_metrics, self.pixel_metrics, outputs)
-
-    def on_validation_epoch_start(self):
-        self.image_metrics.reset()
-        self.pixel_metrics.reset()
-
-    def on_validation_epoch_end(self):
-        self._update_metrics_threshold()
-
-        self._log_metrics()
-
-    def on_test_epoch_start(self):
-        self.image_metrics.reset()
-        self.pixel_metrics.reset()
-
-    def on_test_epoch_end(self):
-        self._log_metrics()
-
-    def _update_metrics_threshold(self) -> None:
-        self.image_metrics.set_threshold(self.image_threshold.value.item())
-        self.pixel_metrics.set_threshold(self.pixel_threshold.value.item())
-
-    @staticmethod
-    def _collect_output(
-        image_metric: AnomalibMetricCollection,
-        pixel_metric: AnomalibMetricCollection,
-        output: STEP_OUTPUT,
-    ) -> None:
-        image_metric.cpu()
-        image_metric.update(output["pred_scores"], output["label"].int())
-        if "mask" in output.keys() and "anomaly_maps" in output.keys():
-            pixel_metric.cpu()
-            pixel_metric.update(torch.squeeze(output["anomaly_maps"]), torch.squeeze(output["mask"].int()))
-
-    def _outputs_to_cpu(self, output):
-        if isinstance(output, dict):
-            for key, value in output.items():
-                output[key] = self._outputs_to_cpu(value)
-        elif isinstance(output, list):
-            output = [self._outputs_to_cpu(item) for item in output]
-        elif isinstance(output, Tensor):
-            output = output.cpu()
-        return output
-
-    def _log_metrics(self) -> None:
-        """Log computed performance metrics."""
-        if self.pixel_metrics._update_called:
-            self.log_dict(self.pixel_metrics, prog_bar=True)
-            self.log_dict(self.image_metrics, prog_bar=False)
-        else:
-            self.log_dict(self.image_metrics, prog_bar=True)
