+++ src/anomalib/trainer/callbacks/__init__.py
@@ -3,8 +3,9 @@
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
+from .metrics_manager import MetricsManagerCallback
 from .normalization import get_normalization_callback
 from .post_processor import PostProcessorCallback
 from .thresholding import ThresholdingCallback
 
-__all__ = ["PostProcessorCallback", "ThresholdingCallback", "get_normalization_callback"]
+__all__ = ["MetricsManagerCallback", "PostProcessorCallback", "ThresholdingCallback", "get_normalization_callback"]
